+++ scoda/sequences/absolute_sequence.py
@@ -52,14 +52,14 @@ class AbsoluteSequence(AbstractSequence):
 
         return True
 
-    def absolute_note_array(self, standard_length=PPQN, include_internal_messages: bool = False) -> [
+    def absolute_note_array(self, standard_length=PPQN, include_meta_messages: bool = False) -> [
         [Message, Message]]:
         """Creates an array containing tuples of messages corresponding to the opening and closing of a note.
 
         Args:
             standard_length: The length used for notes that have not been closed
-            include_internal_messages: Whether internal messages should be added to the array.
-                Internal messages are represented using a tuple of length 1.
+            include_meta_messages: Whether other message types should be added to the array. These messages are
+                represented using a tuple of length 1. Currently only includes internal and time signature messages.
 
         Returns: An array of tuples of two messages constituting a note
 
@@ -93,8 +93,9 @@ class AbsoluteSequence(AbstractSequence):
                     index = open_messages.pop(msg.note)
                     notes[index].append(msg)
 
-            elif msg.message_type == MessageType.INTERNAL and include_internal_messages:
-                notes.insert(i, [Message(message_type=MessageType.INTERNAL, time=msg.time)])
+            elif include_meta_messages and (msg.message_type == MessageType.TIME_SIGNATURE or
+                                            msg.message_type == MessageType.INTERNAL):
+                notes.insert(i, [msg])
                 i += 1
 
         # Check unclosed notes